Долго пытался сделать механику лазера, он должен появляется (предупреждать), а дальше становится опастным. По консоли он появляется но почему-то не там где нужно... Ещё тут таймеров нет, использованы переменные времени

+++ core/src/ru/kaverznev/spacestudio/objects/UFOObject.java
@@ -43,8 +43,7 @@ public class UFOObject extends GameObject {
 
 
     public void update() {
-        // Oscillate horizontally around the center with a controlled amplitude
-        float horizontalMovement = (float) Math.sin(getY() / 20) * 75;
+        float horizontalMovement = (float) Math.sin((double) getY() / 20) * 75;
         body.setLinearVelocity(new Vector2(horizontalMovement, -UFO_VELOCITY));
     }
